refactor: 테스트 픽스쳐 생성.
