+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -1,58 +1,26 @@
 package nextstep.subway.station;
 
+import static nextstep.subway.station.StationFixture.지하철역_목록_조회;
+import static nextstep.subway.station.StationFixture.지하철역_삭제;
+import static nextstep.subway.station.StationFixture.지하철역_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.AcceptanceTest;
-
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
+import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
 
-    public static ExtractableResponse<Response> 지하철역_목록_조회() {
-        return RestAssured.given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/stations")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철역_생성(final String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-
-        return 지하철역_생성(params);
-    }
-
-    public static ExtractableResponse<Response> 지하철역_생성(Map<String, String> params) {
-        return RestAssured.given().log().all()
-            .body(params)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().post("/stations")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철역_삭제(final Long id) {
-        return RestAssured.given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().delete("/stations/" + id)
-            .then().log().all()
-            .extract();
-    }
-
     /**
-     * When 지하철역을 생성하면 Then 지하철역이 생성된다 Then 지하철역 목록 조회 시 생성한 역을 찾을 수 있다
+     * When 지하철역을 생성하면
+     * Then 지하철역이 생성된다
+     * Then 지하철역 목록 조회 시 생성한 역을 찾을 수 있다
      */
     @DisplayName("지하철역을 생성한다.")
     @Test
@@ -73,7 +41,9 @@ public class StationAcceptanceTest extends AcceptanceTest {
     }
 
     /**
-     * Given 지하철역을 생성하고 When 기존에 존재하는 지하철역 이름으로 지하철역을 생성하면 Then 지하철역 생성이 안된다
+     * Given 지하철역을 생성하고
+     * When 기존에 존재하는 지하철역 이름으로 지하철역을 생성하면
+     * Then 지하철역 생성이 안된다
      */
     @DisplayName("기존에 존재하는 지하철역 이름으로 지하철역을 생성한다.")
     @Test
@@ -89,7 +59,9 @@ public class StationAcceptanceTest extends AcceptanceTest {
     }
 
     /**
-     * Given 2개의 지하철역을 생성하고 When 지하철역 목록을 조회하면 Then 2개의 지하철역을 응답 받는다
+     * Given 2개의 지하철역을 생성하고
+     * When 지하철역 목록을 조회하면
+     * Then 2개의 지하철역을 응답 받는다
      */
     @DisplayName("지하철역을 조회한다.")
     @Test
@@ -106,7 +78,9 @@ public class StationAcceptanceTest extends AcceptanceTest {
     }
 
     /**
-     * Given 지하철역을 생성하고 When 그 지하철역을 삭제하면 Then 그 지하철역 목록 조회 시 생성한 역을 찾을 수 없다
+     * Given 지하철역을 생성하고
+     * When 그 지하철역을 삭제하면
+     * Then 그 지하철역 목록 조회 시 생성한 역을 찾을 수 없다
      */
     @DisplayName("지하철역을 제거한다.")
     @Test
@@ -120,4 +94,5 @@ public class StationAcceptanceTest extends AcceptanceTest {
         //then
         assertThat(지하철역_목록_조회().jsonPath().getList("")).isEmpty();
     }
+
 }
